filestream request reader
